dfu: throw exception when vid & pid == 0; change altsetting type

+++ src/dfu.hpp
@@ -25,7 +25,7 @@ class DFU {
 		 * \param[in] verbose_lvl: verbose level 0 normal, -1 quiet, 1 verbose
 		 */
 		DFU(const std::string &filename, uint16_t vid, uint16_t pid,
-				int altsetting, int verbose_lvl);
+				int16_t altsetting, int verbose_lvl);
 
 		~DFU();
